pyproject.toml
[project]
name = "websearchagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.7",
    "streamlit>=1.48.1",
    "tavily-python>=0.7.10",
]
